fixup! fixup! fixup! tests/periph_spi: Add thread runtime stats

+++ tests/periph_spi/main.c
@@ -280,8 +280,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 1 - write %i times %i byte:", BENCH_REDOS, 1);
-    printf("\t\t\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -296,8 +296,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 2 - write %i times %i byte:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t\t\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -312,8 +312,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 3 - write %i times %i byte:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -328,8 +328,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 4 - write %i times %i byte to register:", BENCH_REDOS, 1);
-    printf("\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -344,8 +344,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 5 - write %i times %i byte to register:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -360,8 +360,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 6 - write %i times %i byte to register:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -376,8 +376,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 7 - read %i times %i byte:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t\t\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -392,8 +392,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 8 - read %i times %i byte:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t\t\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -408,8 +408,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 9 - read %i times %i byte from register:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -424,8 +424,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf("10 - read %i times %i byte from register:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -440,8 +440,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf("11 - transfer %i times %i byte:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -456,8 +456,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf("12 - transfer %i times %i byte:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -472,8 +472,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf("13 - transfer %i times %i byte to register:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
@@ -488,8 +488,8 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf("14 - transfer %i times %i byte to register:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t%i\t%i\n", (int)(stop - start),
-           (int)xtimer_diff(sched_stop, sched_start).ticks32);
+    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+           xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
                                                         sched_start));
